Add better support for Kotlin types
